requirements.txt
pyserial>=3.5
numpy>=1.20.0
sounddevice>=0.4.4
scipy>=1.7.0